Make random never output the top of the bound
Conform to the same range regardless of inputs. Changes the single
argument version of random to use the same bounds as the other two.

lume.random(), lume.random(1), lume.random(0,1) should all mean the same
thing: output a number in [0,1). They didn't all produce the same
results. The middle option would produce results in (0,1].

Add a test to try some iterations to test failures. An older version of
this test caught errors for the single argument version, but with lua54
I no longer see those failures (the sequence of numbers of the same seed
differs) so I've reworked the test.

Don't use testeq for each iteration like this:
    testeq(0 <= a, true)
    testeq(a < 1, true)
Because all that printing is very slow.

All tests pass.